Smtlib: don't hang on read() forever if we launch a non-existant proc
With this change, if I set my z3_path to "z3333333", I get:

```
$ ./Main.d.byte tests/file.c
solver child (pid 19717) exited
Fatal error: exception Sys_error("Bad file descriptor")
```

Basically - it causes an exception to be thrown in command, either
when writing to or reading from the file descriptor.  This was the
only non-racy way I could think to do this, but it is certainly ugly
and the error is completely non-obvious.